separated init tests concerning wi4mpi to allow to undercut them if the mpi compiler fails or none is found
